pyproject.toml
[project]
name = "youtube-api-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
